requirements.txt
django==4.2.5
django_rest_framework==0.1.0
python-dotenv==1.0.0
psycopg2==2.9.7
requests==2.31.0